Navweb: multi-language exec, navigation menu, css

+++ gramps/plugins/webreport/download.py
@@ -67,12 +67,12 @@ class DownloadPage(BasePage):
     """
     This class is responsible for displaying information about the Download page
     """
-    def __init__(self, report, title):
+    def __init__(self, report, the_lang, the_title):
         """
         @param: report -- The instance of the main report class for this report
         @param: title  -- Is the title of the web page
         """
-        BasePage.__init__(self, report, title)
+        BasePage.__init__(self, report, the_lang, the_title)
 
         # do NOT include a Download Page
         if not self.report.inc_download:
@@ -140,10 +140,10 @@ class DownloadPage(BasePage):
                                 dwnld += 1
                                 trow = Html("tr", id='Row01')
                                 tbody += trow
-
+                                fname_lnk = "../" + fname if the_lang else fname
                                 dldescrx = dldescr[fnamex]
                                 tcell = Html("td", class_="ColumnFilename") + (
-                                    Html("a", fname, href=fname,
+                                    Html("a", fname, href=fname_lnk,
                                          title=html_escape(dldescrx))
                                 )
                                 trow += tcell
@@ -163,9 +163,10 @@ class DownloadPage(BasePage):
                                     last_mod = datetime.datetime.fromtimestamp(
                                         modified)
                                     tcell += last_mod
-                                    # copy the file
-                                    self.report.copy_file(dlfname[fnamex],
-                                                          fname)
+                                    # copy the file only once
+                                    if not os.path.exists(fname):
+                                        self.report.copy_file(dlfname[fnamex],
+                                                              fname)
                                 else:
                                     tcell += self._("Cannot open file")
